Updated Chapter 1 for Swift 4
